修改商品价格及商品数量bug

+++ src/main/java/com/sysongy/poms/ordergoods/model/SysOrderGoods.java
@@ -11,7 +11,7 @@ public class SysOrderGoods {
 
     private BigDecimal price;
 
-    private Integer number;
+    private Double number;
 
     private BigDecimal sumPrice;
 
@@ -43,11 +43,11 @@ public class SysOrderGoods {
         this.price = price;
     }
 
-    public Integer getNumber() {
+    public Double getNumber() {
         return number;
     }
 
-    public void setNumber(Integer number) {
+    public void setNumber(Double number) {
         this.number = number;
     }
